Refactor _private out into separate Utilities file/global on Raygun obj
