+++ Gruntfile.js
@@ -22,9 +22,9 @@ module.exports = function(grunt) {
       },
       dist: {
         files: {
-          'dist/raygun.js': ['tracekit/tracekit.js', 'src/raygun.tracekit.jquery.js', 'src/raygun.js', 'src/raygun.rum.js', 'src/raygun.js-url.js', 'src/raygun.loader.js'],
-          'dist/raygun.vanilla.js': ['tracekit/tracekit.js', 'src/raygun.js', 'src/raygun.rum.js', 'src/raygun.js-url.js', 'src/raygun.loader.js'],
-          'dist/raygun.umd.js': ['src/umd.intro.js', 'tracekit/tracekit.js', 'src/raygun.tracekit.jquery.js', 'src/raygun.js', 'src/raygun.rum.js', 'src/raygun.persistence.js', 'src/raygun.js-url.js', 'src/raygun.loader.js', 'src/umd.outro.js']
+          'dist/raygun.js': ['tracekit/tracekit.js', 'src/raygun.tracekit.jquery.js', 'src/raygun.utilities.js', 'src/raygun.js', 'src/raygun.rum.js', 'src/raygun.loader.js'],
+          'dist/raygun.vanilla.js': ['tracekit/tracekit.js', 'src/raygun.utilities.js', 'src/raygun.js', 'src/raygun.rum.js', 'src/raygun.loader.js'],
+          'dist/raygun.umd.js': ['src/umd.intro.js', 'tracekit/tracekit.js', 'src/raygun.tracekit.jquery.js', 'src/raygun.utilities.js', 'src/raygun.js', 'src/raygun.rum.js', 'src/raygun.persistence.js', 'src/raygun.loader.js', 'src/umd.outro.js']
         }
       }
     },
